third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-over-across-timeouts.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    setTimeout(callback1, 0);
}

function callback1()
{
    setTimeout(callback2, 0);
    debugger;
}

function callback2()
{
    var dummy = 42; // Should pause here.
    (function FAIL_Should_Not_Pause_Here() { debugger; })();

    if (window.testRunner)
        testRunner.evaluateInWebInspector(0, "(SourcesTestRunner.completeDebuggerTest())")
}

var test = function() {
  SourcesTestRunner.startDebuggerTest(step1, true);

  function step1() {
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(step2);
  }

  function step2() {
    var actions = [
      'Print',  // "debugger" in callback1
      'StepOver',
      'StepOver',
      'Print',
    ];
    SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions, resume);
  }

  function resume() {
    SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(InspectorTest, resume));
  }
};

</script>
</head>

<body onload="runTest()">
<p>
Tests that debugger StepOver will stop inside next timeout handler.
</p>
</body>
</html>
