third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-over-across-timeouts-expected.txt

Tests that debugger StepOver will stop inside next timeout handler.

Set timer for test function.
Call stack:
    0) callback1 (debugger-step-over-across-timeouts.html:15)

Executing StepOver...
Executing StepOver...
Call stack:
    0) callback2 (debugger-step-over-across-timeouts.html:18)
